+++ Sources/ArcGISToolkit/Components/Offline/Utilities/FileManager+Offline.swift
@@ -30,36 +30,4 @@ extension FileManager {
         }
         return accumulatedSize
     }
-    
-    /// Returns a Boolean value indicating if the specified directory is empty.
-    /// - Parameter path: The path to check.
-    func isDirectoryEmpty(atPath path: URL) -> Bool {
-        (try? FileManager.default.contentsOfDirectory(atPath: path.path()).isEmpty) ?? true
-    }
-}
-
-extension URL {
-    /// The path to the web map directory for a specific portal item.
-    /// `Documents/OfflineMapAreas/<Portal Item ID>/`
-    /// - Parameter portalItemID: The ID of the web map portal item.
-    private static func portalItemDirectory(forPortalItemID portalItemID: PortalItem.ID) -> URL {
-        URL.documentsDirectory.appending(components: "OfflineMapAreas", "\(portalItemID)/")
-    }
-    
-    /// The path to the directory for a specific map area from the preplanned map areas directory for a specific portal item.
-    /// `Documents/OfflineMapAreas/<Portal Item ID>/Preplanned/<Preplanned Area ID>/`
-    /// - Parameters:
-    ///   - portalItemID: The ID of the web map portal item.
-    ///   - preplannedMapAreaID: The ID of the preplanned map area portal item.
-    /// - Returns: A URL to the preplanned map area directory.
-    static func preplannedDirectory(
-        forPortalItemID portalItemID: PortalItem.ID,
-        preplannedMapAreaID: PortalItem.ID? = nil
-    ) -> URL {
-        if let preplannedMapAreaID {
-            portalItemDirectory(forPortalItemID: portalItemID).appending(components: "Preplanned", "\(preplannedMapAreaID)/")
-        } else {
-            portalItemDirectory(forPortalItemID: portalItemID).appending(components: "Preplanned/")
-        }
-    }
 }
